backend/requirements.txt
fastapi==0.115.12
pydantic==2.11.3
uvicorn==0.34.2
openai==1.76.0
python-dotenv==1.1.0
aioredis==2.0.1
asyncio==3.4.3